requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.3
django-crispy-forms==1.14.0
djangorestframework==3.13.1
Pillow==9.0.1
pytz==2022.1
sqlparse==0.4.2
